requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
itsdangerous==2.1.2
redis==5.0.1
aiofiles==23.2.1